Admin condition to access pages

+++ controllers/userController.php
@@ -4,7 +4,17 @@ $action = $_REQUEST['action'];
 
 switch ($action)
 {
+    case 'home':
+        if(isset($_SESSION['admin'])){
+            $users = $pdo->getUsers();
+            include("view/home.php");
+        }else{
+            header("Location: connection");
+        }
+    break;
+
     case 'control': //le case cest la valeur attribuer a Action=..
+        
         $email = $_REQUEST['email'];
         $password = $_REQUEST['password'];
         $user = $pdo->login($email, $password);
@@ -27,43 +37,47 @@ switch ($action)
     break;
     
     case 'createUser':
-        $email = $_REQUEST['email'];
-        $password = $_REQUEST['password'];
-        $firstName = $_REQUEST['firstName'];
-        $lastName = $_REQUEST['lastName'];
-        $adress = $_REQUEST['adress'];
-        $postalCode = $_REQUEST['postalCode'];
-        $city = $_REQUEST['city'];
-        $user = $pdo->createUser($email, $password, $firstName, $lastName, $adress, $postalCode, $city);
+        if(isset($_SESSION['admin'])){
+            $email = $_REQUEST['email'];
+            $password = $_REQUEST['password'];
+            $firstName = $_REQUEST['firstName'];
+            $lastName = $_REQUEST['lastName'];
+            $adress = $_REQUEST['adress'];
+            $postalCode = $_REQUEST['postalCode'];
+            $city = $_REQUEST['city'];
+            $user = $pdo->createUser($email, $password, $firstName, $lastName, $adress, $postalCode, $city);
         header("Location: home");
-    break;
-    
-    case 'home':
-        $users = $pdo->getUsers();
-        include("view/home.php");
+        }else{
+            header("Location: connection");
+        }
     break;
 
     case 'delete':
-        
-        $id = intval($_GET['id']);
-        $delete = $pdo->deleteUser($id);
-        header("Location:home");
-        // include("view/home.php");
+        if(isset($_SESSION['admin'])){
+            $id = intval($_GET['id']);
+            $delete = $pdo->deleteUser($id);
+            header("Location:home");
+            // include("view/home.php");
+        }else{
+            header("Location: connection");
+        }
     break;
     
     case 'update':
-        
-        $id = intval($_GET['id']);
-        $firstName = $_REQUEST['firstName'];
-        $lastName = $_REQUEST['lastName'];
-        $adress = $_REQUEST['adress'];
-        $postalCode = $_REQUEST['postalCode'];
-        $city = $_REQUEST['city'];
-        
-        $res = $pdo->updateUser($firstName, $lastName, $adress, $postalCode, $city, $id);
+        if(isset($_SESSION['admin'])){
+            $id = intval($_GET['id']);
+            $firstName = $_REQUEST['firstName'];
+            $lastName = $_REQUEST['lastName'];
+            $adress = $_REQUEST['adress'];
+            $postalCode = $_REQUEST['postalCode'];
+            $city = $_REQUEST['city'];
+            
+            $res = $pdo->updateUser($firstName, $lastName, $adress, $postalCode, $city, $id);
 
-        header("Location: home");
-
+            header("Location: home");
+        }else{
+            header("Location: connection");
+        }
         break;
         
     default:
